refactor: 도메인, 메소드 명 변경 (#411)
