+++ src/main/java/com/festival/domain/viewcount/model/Home.java
@@ -1,4 +1,4 @@
-package com.festival.domain.viewcount;
+package com.festival.domain.viewcount.model;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
