+++ src/main/java/com/festival/domain/viewcount/repository/HomeRepository.java
@@ -1,6 +1,6 @@
 package com.festival.domain.viewcount.repository;
 
-import com.festival.domain.viewcount.Home;
+import com.festival.domain.viewcount.model.Home;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface HomeRepository extends JpaRepository<Home, Long> {
