+++ src/main/java/com/festival/domain/viewcount/service/HomeService.java
@@ -1,6 +1,8 @@
 package com.festival.domain.viewcount.service;
 
-import com.festival.domain.viewcount.Home;
+import com.festival.common.exception.ErrorCode;
+import com.festival.common.exception.custom_exception.NotFoundException;
+import com.festival.domain.viewcount.model.Home;
 import com.festival.domain.viewcount.repository.HomeRepository;
 import com.festival.domain.viewcount.util.ViewCountUtil;
 import lombok.RequiredArgsConstructor;
@@ -8,13 +10,14 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 @Service
 public class HomeService {
 
     private final ViewCountUtil viewCountUtil;
     private final HomeRepository homeRepository;
 
-    public void increase(String ipAddress) {
+    public void increaseViewCount(String ipAddress) {
         if(!viewCountUtil.isDuplicatedAccess(ipAddress, "Home")) {
             viewCountUtil.increaseData("viewCount_Home_1");
             viewCountUtil.setDuplicateAccess(ipAddress, "Home");
@@ -22,8 +25,8 @@ public class HomeService {
     }
 
     @Transactional
-    public void update(Long id, long count){
-        Home home = homeRepository.findById(id).orElse(null);
+    public void updateViewCount(Long id, long count){
+        Home home = homeRepository.findById(id).orElseThrow(() -> new NotFoundException(ErrorCode.NOT_FOUND_OBJECT));
         home.plus(count);
     }
 }
